LayoutTests/fast/mediastream/now-playing-and-mediastream.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id=video autoplay playsinline></video>
        <script>
function waitFor(delay)
{
    return new Promise(resolve => setTimeout(resolve, delay));
}

async function waitForCriteria(test, criteria)
{
    let counter = 0;
    while (!criteria() && ++counter < 100)
        await waitFor(50);
}

promise_test(async test => {
    if (!window.internals)
        return;

    assert_false(!!internals.nowPlayingState.uniqueIdentifier, "test1");

    video.srcObject = await navigator.mediaDevices.getUserMedia({ audio:true, video:true });
    await video.play();

    await waitFor(500);
    assert_false(!!internals.nowPlayingState.uniqueIdentifier, "test2");

    navigator.mediaSession.metadata = new MediaMetadata({ title: "test" });
    await waitForCriteria(test, () => !!internals.nowPlayingState.uniqueIdentifier);
    assert_true(!!internals.nowPlayingState.uniqueIdentifier, "test3");

    navigator.mediaSession.metadata = null;
    await waitForCriteria(test, () => !internals.nowPlayingState.uniqueIdentifier);
    assert_false(!!internals.nowPlayingState.uniqueIdentifier, "test4");

    navigator.mediaSession.setActionHandler("pause", () => { });
    await waitForCriteria(test, () => !!internals.nowPlayingState.uniqueIdentifier);
    assert_true(!!internals.nowPlayingState.uniqueIdentifier, "test5");

    navigator.mediaSession.setActionHandler("pause", null);
    await waitForCriteria(test, () => !internals.nowPlayingState.uniqueIdentifier);
    assert_false(!!internals.nowPlayingState.uniqueIdentifier, "test6");
}, "HTMLMediaElement will become the now playing info source if using mediaSession API");

        </script>
    </body>
</html>
